Removed hard coded window and hop sizes. Fixed vocoder intialization.
The hard coded windown and hop sizes were causing the reconstruction to sound
flubby. At the moment a window is only applied on the resynthesis. That's
because the current models were trained with a rectangular window on the front
end. This is an error. Future models should have hanning window on analysis and
rectangular window on resynthesis

When the vocoder is turned on, the intial phase is now random. This has helped iin removing clicks